omxbufferpool: Do not handle GstOMXBuffer in freeing proccess
GstOMXBuffers are created in omxdecoder/encoder. They will be
deallocated in its freeing process. This freeing process is separated
with freeing process of omxbufferpool.
Do not handle GstOMXBuffer in freeing process of omxbufferpool to avoid
not synchronization between freeing in omxdecoder/encoder and omxbufferpool
(GstOMXBuffer can be deallocated before omxbufferpool finish freeing) and
lead to "Caught SIGSEGV" issue

Currently, dma_id is contained in GstOMXBuffer and it is used
to end exporting dmabuf in freeing process of omxbufferpool.
In case downstream is waylandsink, freeing of omxbufferpool must wait for
unref GstBuffer from waylandsink.
It lead to not synchronization and cause "Caught SIGSEGV".
This patch file will contain dma_id in omxbufferpool throught an GArray
instead of GstOMXBuffer to avoid issue "Caught SIGSEGV".

refs #78736
Signed-off-by: Thao Luong <thao.luong.uw@rvc.renesas.com>