omxbufferpool: Correct allocator before allocate buffer
There are 2 allocators can be used when allocate memory for buffer: dmabuf and
openmax allocators.
dmabuf allocator will be used to create for dmabuf memory only however there will
be case which use dmabuf allocator to create for OMX memory (decode for video
has 1 or 2 frame in case dmabuf mode).
This lead to freeing for memory after using is not correct anymore
(Use freeing of dmabuf allocator for OMX memory)

This patch file will check allocator before create buffer, correct it if it is wrong.

refs #104536
Signed-off-by: Thao Luong <thao.luong.uw@rvc.renesas.com>