omxvideodec: Fix issue when caps is updated 2 times
When caps is updated on the second time, video can not decode.
Reason: Lack of handling for reconfigure flow on set_format()
Solution: Update handling for reconfigure on set_format():
- Add condition to set parameter for OMXR header (headers
support for no-reorder and lossy-compress) only on the first
time update caps to avoid seting parameter when OMX is StateExecuting
when reconfigure.
- Create buffer pool for output port in case reconfigure
- Populate for output port

refs #104238
Signed-off-by: Thao Luong <thao.luong.uw@rvc.renesas.com>