omxh264enc: Update get_caps to handle for case: OMX support auto judgment for level
OMX component supports auto judgment for level base on parameters
(such as resolution, ...) to avoid violation standard in bitstream.
But this update can lead to not-negotiated issue on GStreamer
because the actual level set on caps will be different from setting of user on pipeline.

This patch file will check level is set on pipeline and actual level return from OMX.
If there is difference, and it is under auto judgment mode, still keep level is set on
caps as setting of user on pipeline but will notify to user about the actual level in
bitstream via warning message.

refs #112698
Signed-off-by: Thao Luong <thao.luong.uw@rvc.renesas.com>