omxvideodec: Add use-dmabuf option for omx video decoders
This option will share dmabuf_fd of OMXBuffer with downstream.

mmngrbuf is used to export dmabuf_fd of decoded data (OMXBuffer)
So this option will be enable as default when mmngrbuf is supported.
Handling for creating buffer contain dmabuf_fd will be done on
omxbufferpool which will be updated in next commit.
This patch file only add use-dmabuf option to video decoder.

Can disable support of mmngrbuf as well as dmabuf by setting
"--disable-dmabuf" in built-in configure.

refs #138391
Signed-off-by: Anh Ha <anh.ha.wz@rvc.renesas.com>