omxvideoenc: implement dmabuf export on input buffers
Propose pool upstream so input buffers can be allocated by the port and
exported as dmabuf.

The actual OMX buffers are allocated when the pool is activated, so we
don't end up doing useless allocations if the pool isn't used.

https://bugzilla.gnome.org/show_bug.cgi?id=796918

gst_omx_port_set_dmabuf () is not called
because it is not necessary for RCar and it is not backported.

(cherry picked from commit 7be54ad091fc185c92ebdb60d470e26a17938779)
Signed-off-by: Shinya Saito <ssaito@igel.co.jp>