omxbufferpool: reference the OMX component
Now that the pool is responsible of freeing the OMX buffers, we need to
ensure that the OMX component stay alive while the pool is as we rely on
the component to free the buffers.
The GstOMXPort is owned by the component so no need to ref this one.

https://bugzilla.gnome.org/show_bug.cgi?id=796918
(cherry picked from commit 86a6703d918e932ef9bd3d0618fd7d5e52e3a94d)