omxvideodec: Only update crop info on reconfigure
The crop region can only be reliably updated as part of the
port reconfiguration procedure.  Othewise, there is no
way to know which decoded buffers correspond to the which
crop region (pre or post change).

The crop region is available when the first frame is emitted
from the component after reconfiguration is complete.

Only query the new crop region during the first frame after
reconfiguration and keep a copy of the result in the decoder
context.